refactor: MyPage 컴포넌트를 components/common/에서 components/my/로 분리
