+++ src/app/my/account-info/page.jsx
@@ -3,9 +3,9 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { jwtDecode } from 'jwt-decode';
 import useTokenStore from '../../../stores/useTokenStore';
-import axiosInstance from '../../../libs/api/instance';
-import MyPageHeader from '@components/common/MyPageHeader';
-import MyPageBlock from '@components/common/MyPageBlock';
+import axiosInstance from '@lib/api/instance';
+import MyPageHeader from '@components/my/MyPageHeader';
+import MyPageBlock from '@components/my/MyPageBlock';
 import Modal from '@components/common/Modal';
 import LoginRequiredModal from '@components/common/LoginRequiredModal';
 import PrivacyPolicyFooter from '@components/common/PrivacyPolicyFooter';
