+++ src/app/my/cancel-account-step1/page.jsx
@@ -4,10 +4,10 @@ import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import { jwtDecode } from 'jwt-decode';
 import useTokenStore from '../../../stores/useTokenStore';
-import axiosInstance from '../../../libs/api/instance';
+import axiosInstance from '@lib/api/instance';
 import Button from '../../../components/common/Button';
 import PrivacyPolicyFooter from '../../../components/common/PrivacyPolicyFooter';
-import MyPageHeader from '@components/common/MyPageHeader';
+import MyPageHeader from '@components/my/MyPageHeader';
 import Modal from '@components/common/Modal';
 import LoginRequiredModal from '@components/common/LoginRequiredModal';
 import FooterNav from '../../../components/common/FooterNav';
